Fold PersonFeedbackService into PersonService
The feedback service was accidentally split out of FaceRecognitionService
and only orchestrated PersonService state. PersonService.syncPersonFeedback
is now the single serialized entry point; the pull-and-apply and reconcile
primitives are private, so nothing can bypass the serialization. Clustering
now runs a pending reconcile before snapshotting persons.
