--- mobile/apps/photos/lib/services/machine_learning/face_ml/person/person_feedback_service.dart
@@ -1,68 +1,0 @@
-import "dart:async" show unawaited;
-
-import "package:logging/logging.dart";
-import "package:photos/core/event_bus.dart";
-import "package:photos/events/diff_sync_complete_event.dart";
-import "package:photos/events/people_changed_event.dart";
-import "package:photos/service_locator.dart" show isLocalGalleryMode;
-import "package:photos/services/machine_learning/face_ml/person/person_service.dart";
-
-class PersonFeedbackService {
-  static final _logger = Logger("PersonFeedbackService");
-
-  PersonFeedbackService() {
-    Bus.instance.on<DiffSyncCompleteEvent>().listen((event) {
-      unawaited(syncPersonFeedback());
-    });
-    Bus.instance.on<PeopleChangedEvent>().listen((event) {
-      if (event.type != PeopleEventType.syncDone) {
-        _shouldReconcilePeople = true;
-      }
-    });
-  }
-
-  bool _shouldReconcilePeople = false;
-  bool _syncRequested = false;
-  Future<void>? _syncFuture;
-
-  Future<void> syncPersonFeedback() {
-    _syncRequested = true;
-    return _syncFuture ??= _runPendingSyncs();
-  }
-
-  Future<void> _runPendingSyncs() async {
-    try {
-      do {
-        _syncRequested = false;
-        await _syncPersonFeedbackInternal();
-      } while (_syncRequested);
-    } finally {
-      _syncFuture = null;
-    }
-  }
-
-  Future<void> _syncPersonFeedbackInternal() async {
-    if (isLocalGalleryMode) {
-      _logger.finest("Skipping person feedback sync in local gallery mode");
-      return;
-    }
-
-    if (_shouldReconcilePeople) {
-      _shouldReconcilePeople = false;
-      try {
-        await PersonService.instance.reconcileClusters();
-      } catch (_) {
-        _shouldReconcilePeople = true;
-        rethrow;
-      }
-      Bus.instance.fire(PeopleChangedEvent(type: PeopleEventType.syncDone));
-    } else {
-      final didChange = await PersonService.instance
-          .fetchRemoteClusterFeedback();
-      if (didChange) {
-        _logger.info("people: got remote data update");
-        Bus.instance.fire(PeopleChangedEvent(type: PeopleEventType.syncDone));
-      }
-    }
-  }
-}
