+++ mobile/apps/photos/lib/services/machine_learning/ml_service.dart
@@ -352,7 +352,7 @@ class MLService {
 
   Future<void> sync() async {
     await fileDataService.syncFDStatus();
-    await personFeedbackService.syncPersonFeedback();
+    await PersonService.instance.syncPersonFeedback();
   }
 
   Future<void> runAllML({bool force = false}) async {
@@ -569,7 +569,7 @@ class MLService {
     final faceIdNotToCluster = <String, List<String>>{};
     if (!isLocalGalleryMode) {
       _logger.info('Pulling remote feedback before actually clustering');
-      await PersonService.instance.fetchRemoteClusterFeedback();
+      await PersonService.instance.syncPersonFeedback();
       final persons = await PersonService.instance.getPersons();
       for (final person in persons) {
         if (person.data.rejectedFaceIDs.isNotEmpty) {
